Added Filling to all Shapes
Big update- most addNewObjects functions now require a boolean saying if filled or not and the colour.The fill is displayed while dragging all components that can be filled (besides polygon). Added in a helperClass for the polygons that keeps the same index as the other polygon list and contains info on the if the poly is filled, and the colours of border and fill. Added some more tests for the filling.