Added ability to undo last object drawn
Created linkedlist with the component name as a string and the index of the individual object in the array (this will be used later with the ability to remove any drawing previously added and list all the drawings done). Each class has a new function which removes 1 entry at the index specified. Also reworked the polygon drawing to be consistent with the way the other shapes are drawn. Some tests have been to test the new functions.