Added scaling (all values are now floats) and enum of all shapes
All shapes take float values which acts a % of the screen to be filled up by the image. Polygon has been refactored(again 🙄) and now doesnt store the polygon in its list rather the array of points, the polygon is then generated at the draw method. All graphics are now done with 2DGraphics. The enum of shapes is for consistency.